+++ src/mefit/Common/Constants/ApplicationChars.cs
@@ -1,0 +1,20 @@
+﻿// Mac EFI Toolkit
+// https://github.com/MuertoGB/MacEfiToolkit
+
+// ApplicationChars.cs
+// Released under the GNU GLP v3.0
+
+namespace Mac_EFI_Toolkit.Common.Constants
+{
+    internal class ApplicationChars
+    {
+        internal const string FLUENT_MULTIPLY = "\uE947";
+        internal const string FLUENT_OPENFOLDERHORIZ = "\uED25";
+        internal const string FLUENT_OPENLOCAL = "\uE8E5";
+        internal const string FLUENT_REPORT = "\uE9F9";
+        internal const string FLUENT_OTHERUSER = "\uE7EE";
+        internal const string SEGUI_RIGHTWARDSARROW = "\u2192";
+        internal const string SEGUI_DINGBAT1 = "\u2776";
+        internal const string SEGUI_NOBREAKSPACE = "\u00A0";
+    }
+}